Task TK-93152. Limited the width of the lifespan.
When the width of the renderer container is limited and you have a lifespan with a birth year and "Deceased", the lifespan may end up wrapping. In the "Show Parents" area this results in the parents information being cut off under the wife's information. The lifespan wrapping is most likely to happen in IE and when the language is set to Russian.

It seems best to limit the width of the lifespan and not the width of the PID.